requirements.txt
asgiref==3.10.0
certifi==2025.11.12
charset-normalizer==3.4.4
Django==5.2.7
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework-gis==1.2.0
GDAL==3.11.4
idna==3.11
psycopg2-binary==2.9.11
python-dotenv==1.1.1
requests==2.32.5
sqlparse==0.5.3
urllib3==2.5.0
